requirements.txt
# Core AI / LLM
langchain>=0.2.0
langchain-anthropic>=0.1.15
anthropic>=0.28.0

# FastAPI backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
bcrypt>=4.0.1

# Testing
pytest>=8.0.0
